+++ src/components/SettingsPanel.vue
@@ -40,7 +40,6 @@
 
         <div class="choice-panel-body" :style="{ height: panelHeight + 'px' }">
           <GenerationSettings v-if="activeTab === 'generation'" />
-          <ApiEditor v-else-if="activeTab === 'api'" />
           <WorldInfoEditor v-else-if="activeTab === 'worldinfo'" />
           <FilterEditor v-else-if="activeTab === 'filter'" />
           <AppearanceSettings v-else-if="activeTab === 'appearance'" />
@@ -58,7 +57,6 @@
 
 <script setup lang="ts">
 import { useGlobalSettingsStore } from '@/store/global-settings';
-import ApiEditor from '@/components/ApiEditor.vue';
 import AppearanceSettings from '@/components/AppearanceSettings.vue';
 import GenerationSettings from '@/components/GenerationSettings.vue';
 import FilterEditor from '@/components/FilterEditor.vue';